Move SIgnUp action to BackendDirector from SignWidget
